move create table scripts
